+++ src/utils.cpp
@@ -79,14 +79,14 @@ SDL_Window* createWindow(const char* caption, int width, int height )
 void launchLoop(Application* app)
 {
 	SDL_Event sdlEvent;
-	double last_time = SDL_GetTicks();
+	Float last_time = SDL_GetTicks();
 	int x,y;
 
 	SDL_GetMouseState(&x,&y);
 	app->mouse_position.set(x,y);
 
 #ifdef APP_TIME_ENABLED
-	double start_time = SDL_GetTicks();
+	Float start_time = SDL_GetTicks();
 #endif //APP_TIME_ENABLED
 
 	//infinite loop
@@ -129,8 +129,8 @@ void launchLoop(Application* app)
 		app->mouse_position.set(x,y);
 
 		//update logic
-		double now = SDL_GetTicks();
-		double elapsed_time = (now - last_time) * 0.001; //0.001 converts from milliseconds to seconds
+		Float now = SDL_GetTicks();
+		Float elapsed_time = (now - last_time) * 0.001; //0.001 converts from milliseconds to seconds
 #ifdef APP_TIME_ENABLED
 		app->time = (now - start_time) * 0.001;
 #endif //APP_TIME_ENABLED
